requirements.txt
pandas
requests
feedparser
python-dotenv
scikit-learn
joblib
matplotlib
seaborn
